core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Core frame capture and streaming utilities for Casper agents"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
    "imageio[ffmpeg]>=2.34",
    "livekit>=1.0",
    "opencv-python>=4.13.0.92",
    "python-dotenv>=1.0",
    "ultralytics>=8.4.30",
    "api",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/core"]

[tool.uv.sources]
api = { workspace = true }
